requirements.txt
Django==2.1
djangorestframework==3.8.2
psycopg2==2.8.2
pytz==2018.5
six==1.12.0
cent==3.0.1
Pillow==6.0.0